transformer_model/requirements.txt
flask>=3.0,<4.0
numpy>=1.24
python-dotenv>=1.0
requests>=2.31
Pillow>=10.0
opencv-python>=4.8
